requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0